[REFACTOR] 번들링 최적화 - 소켓 범위 축소
- 소켓 범위를 페이지 전체에서 NavBar로 축소
- (with-socket)도 -> (nav-bar)로 이름 변경
- 아예 밖으로 뺐다가 NavBar 범위도 RootLayout에서 더 내부로 줄이고자 이름만 바뀐 결과가 됨
- 이에 따른 일부 CSS 수정
- SocketContainer도 필요 없어짐
- 경로가 너무 많이 바뀌어서 rm -rf node_modules하면서 다시 설치하니 sockjs-client가 필요하다는 걸 알게 됨
- 재설치
