+++ kakaobase/package.json
@@ -16,6 +16,7 @@
     "@next/bundle-analyzer": "^15.5.0",
     "@stomp/stompjs": "^7.1.1",
     "@tanstack/react-query": "^5.56.2",
+    "@types/sockjs-client": "^1.5.4",
     "axios": "^1.8.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -28,6 +29,7 @@
     "react-hook-form": "^7.53.0",
     "react-linkify": "^1.0.0-alpha",
     "react-use": "^17.5.1",
+    "sockjs-client": "^1.6.1",
     "tailwind-merge": "^3.2.0",
     "tailwind-scrollbar-hide": "^2.0.0",
     "tailwindcss-animate": "^1.0.7",
