+++ kakaobase/src/app/(nav-bar)/main/page.tsx
@@ -1,5 +1,4 @@
 import HeaderMain from '@/widgets/header/HeaderMain';
-import NavBar from '@/widgets/navbar/NavBar';
 import ListRouter from '@/features/feeds/ui/ListRouter';
 import PostCourseSelector from '@/features/feeds/posts/ui/PostCourseSelector';
 import FloatingButtonGroup from '@/widgets/buttons/FloatingButtonGroup';
@@ -13,7 +12,6 @@ export default function Page() {
         <ListRouter />
       </div>
       <FloatingButtonGroup />
-      <NavBar />
     </main>
   );
 }
