+++ kakaobase/src/app/(nav-bar)/page.tsx
@@ -1,7 +1,6 @@
 import Banner from '@/features/main/ui/Banner';
 import PostByCourse from '@/features/main/ui/PostByCourse';
 import HeaderMain from '@/widgets/header/HeaderMain';
-import NavBar from '@/widgets/navbar/NavBar';
 
 export default function Home() {
   return (
@@ -11,7 +10,6 @@ export default function Home() {
         <Banner />
         <PostByCourse />
       </div>
-      <NavBar />
     </div>
   );
 }
