+++ kakaobase/src/app/(nav-bar)/profile/[userId]/page.tsx
@@ -1,5 +1,4 @@
 import Header from '@/widgets/header/Header';
-import NavBar from '@/widgets/navbar/NavBar';
 import MyPageWrapper from '@/features/account/ui/MyPageWrapper';
 
 export default function Page({ params }: { params: { userId: number } }) {
@@ -7,7 +6,6 @@ export default function Page({ params }: { params: { userId: number } }) {
     <div className="flex flex-col h-screen">
       <Header label="프로필" />
       <MyPageWrapper userId={params.userId} />
-      <NavBar />
     </div>
   );
 }
