+++ kakaobase/src/app/(nav-bar)/profile/edit/page.tsx
@@ -1,5 +1,4 @@
 import Header from '@/widgets/header/Header';
-import NavBar from '@/widgets/navbar/NavBar';
 import EditWrapper from '@/features/account/ui/edit/EditWrapper';
 
 export default function Page() {
@@ -9,7 +8,6 @@ export default function Page() {
       <div className="flex-grow flex flex-col h-screen items-center justify-center w-full">
         <EditWrapper />
       </div>
-      <NavBar />
     </div>
   );
 }
