+++ kakaobase/src/app/(nav-bar)/profile/edit/password/page.tsx
@@ -1,5 +1,4 @@
 import Header from '@/widgets/header/Header';
-import NavBar from '@/widgets/navbar/NavBar';
 import PasswordEditWrapper from '@/features/account/ui/password/PasswordEditWrapper';
 
 export default function Page() {
@@ -7,7 +6,6 @@ export default function Page() {
     <div className="flex flex-col h-screen">
       <Header label="비밀번호 변경" />
       <PasswordEditWrapper />
-      <NavBar />
     </div>
   );
 }
