+++ kakaobase/src/app/(nav-bar)/profile/edit/withdraw/page.tsx
@@ -1,5 +1,4 @@
 import Header from '@/widgets/header/Header';
-import NavBar from '@/widgets/navbar/NavBar';
 import WithdrawWrapper from '@/features/account/ui/withdraw/WithdrawWrapper';
 
 export default function Page() {
@@ -7,7 +6,6 @@ export default function Page() {
     <div className="flex flex-col h-screen">
       <Header label="회원 탈퇴" />
       <WithdrawWrapper />
-      <NavBar />
     </div>
   );
 }
