--- kakaobase/src/app/(with-socket)/layout.tsx
@@ -1,14 +1,0 @@
-import SocketContainer from '@/features/socket/container/SocketContainer';
-
-export default async function Layout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <>
-      <SocketContainer />
-      {children}
-    </>
-  );
-}
