+++ kakaobase/src/features/feeds/comments/ui/RecommentStateBar.tsx
@@ -8,7 +8,7 @@ export default function RecommentStateBar() {
   if (!isWritingRecomment) return null;
 
   return (
-    <div className="sticky mb-[4.25rem] px-4 py-2 w-full bg-textColor opacity-40 text-containerColor flex justify-between items-center text-sm">
+    <div className="sticky px-4 py-2 w-full bg-textColor opacity-40 text-containerColor flex justify-between items-center text-sm">
       <div>{commentWriter}에게 남기는 대댓글</div>
       <div
         onClick={stopRecomment}
