--- kakaobase/src/features/socket/container/SocketContainer.tsx
@@ -1,7 +1,0 @@
-'use client';
-import useSocket from '@/features/socket/hooks/useSocket';
-
-export default function SocketContainer() {
-  useSocket();
-  return null;
-}
