+++ kakaobase/src/widgets/navbar/AlarmBadgeClient.tsx
@@ -1,0 +1,10 @@
+import useSocket from '@/features/socket/hooks/useSocket';
+
+export default function AlarmBadgeClient({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  useSocket();
+  return <>{children}</>;
+}
